+++ docker/resources/infrastructure/localstack/localstack-init.sh
@@ -1,2 +1,9 @@
-# At the time of writing (Mar 2020) there are problems using cloudformation templates to initialise SQS queues (FIFO support)
+#!/bin/bash
+###############################################################################
+#  Localstack initialisation
+###############################################################################
+set -e
+trap 'exit 1' INT
+
+# Build a new cloudformation stack to initialise any resources in localstack-cfn.yml
 awslocal cloudformation create-stack --template-body file:///docker-entrypoint-initaws.d/localstack-cfn.yml --stack-name localdev