+++ packages/gafl-webapp-service/src/pages/summary/licence-summary/__tests__/licence-summary.spec.js
@@ -35,9 +35,17 @@ const dobHelper = d => ({
 })
 const dobAdultToday = moment().subtract(JUNIOR_MAX_AGE + 1, 'years')
 
+beforeAll(() => {
+  process.env.ANALYTICS_PRIMARY_PROPERTY = 'UA-123456789-0'
+  process.env.ANALYTICS_XGOV_PROPERTY = 'UA-987654321-0'
+})
 beforeAll(d => start(d))
 beforeAll(d => initialize(d))
 afterAll(d => stop(d))
+afterAll(() => {
+  delete process.env.ANALYTICS_PRIMARY_PROPERTY
+  delete process.env.ANALYTICS_XGOV_PROPERTY
+})
 
 describe('The licence summary page', () => {
   it('redirects to the licence length page if length is set', async () => {