+++ packages/pocl-job/src/pocl-processor.js
@@ -1,3 +1,4 @@
+import config from './config.js'
 import { ftpToS3 } from './transport/ftp-to-s3.js'
 import { s3ToLocal } from './transport/s3-to-local.js'
 import { getFileRecords } from './io/db.js'
@@ -10,6 +11,7 @@ const debug = db('pocl:processor')
 
 export async function execute () {
   try {
+    await config.initialise()
     debug('Retrieving files from FTP')
     await ftpToS3()
     const pendingFileRecords = await getFileRecords(FILE_STAGE.Pending, FILE_STAGE.Staging, FILE_STAGE.Finalising)